kaggle_gemini_data_agent/kaggle_gemini_data_agent/requirements.txt
kaggle>=1.5.12
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.64.0
pandas>=1.3.0
